backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyodbc
pydantic>=2.0
pydantic-settings>=2.0
langchain-core
langchain-groq